+++ src/plone/restapi/tests/test_services_vocabularies.py
@@ -6,6 +6,7 @@ from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.app.testing import TEST_USER_ID
 from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
 from plone.restapi.testing import RelativeSession
+from zope.component import getGlobalSiteManager
 from zope.component import provideUtility
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleVocabulary
@@ -61,6 +62,7 @@ class TestVocabularyEndpoint(unittest.TestCase):
         response = response.json()
         self.assertEqual(
             response,
+<<<<<<< HEAD
             {
                 u'@id': u'http://localhost:55001/plone/@vocabularies/plone.restapi.tests.test_vocabulary',  # noqa
                 u'items': [
@@ -115,6 +117,16 @@ class TestVocabularyEndpoint(unittest.TestCase):
                 u'items_total': 1,
             }
         )
+=======
+            {u'@id': self.portal_url + u'/@vocabularies/''plone.restapi.tests.test_vocabulary',  # noqa
+             u'terms': [
+                 {u'@id': self.portal_url + u'/@vocabularies/plone.restapi.tests.test_vocabulary/token1',  # noqa
+                  u'title': u'Title 1',
+                  u'token': u'token1'},
+                 {u'@id': self.portal_url + u'/@vocabularies/plone.restapi.tests.test_vocabulary/token2',  # noqa
+                  u'title': u'Title 2',
+                  u'token': u'token2'}]})
+>>>>>>> master
 
     def test_get_unknown_vocabulary(self):
         response = self.api_session.get(
@@ -139,7 +151,7 @@ class TestVocabularyEndpoint(unittest.TestCase):
         self.assertEqual(
             [
                 {
-                    u'@id': u'http://localhost:55001/plone/@vocabularies/plone.restapi.tests.test_vocabulary',  # noqa
+                    u'@id': self.portal_url + u'/@vocabularies/plone.restapi.tests.test_vocabulary',  # noqa
                     u'title': u'plone.restapi.tests.test_vocabulary'
                 }
             ],
@@ -166,14 +178,33 @@ class TestVocabularyEndpoint(unittest.TestCase):
         response = self.api_session.get(
             'testdoc/@vocabularies/{}'.format(context_vocab_name))
 
+        gsm = getGlobalSiteManager()
+        gsm.unregisterUtility(provided=IVocabularyFactory,
+                              name=context_vocab_name)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(
             response.json(),
             {
+<<<<<<< HEAD
                 u'@id': u'http://localhost:55001/plone/testdoc/@vocabularies/plone.restapi.tests.test_context_vocabulary',  # noqa
                 u'items': [
                     {u'title': u'testdoc', u'token': u'id'},
                     {u'title': u'Document 1', u'token': u'title'},
                 ],
                 u'items_total': 2,
+=======
+                u'@id': self.portal_url + u'/testdoc/@vocabularies/plone.restapi.tests.test_context_vocabulary',  # noqa
+                u'terms': [
+                    {u'@id': self.portal_url + u'/testdoc/@vocabularies/plone.restapi.tests.test_context_vocabulary/id',  # noqa
+                     u'title': u'testdoc',
+                     u'token': u'id'},
+                    {u'@id': self.portal_url + u'/testdoc/@vocabularies/plone.restapi.tests.test_context_vocabulary/title',  # noqa
+                     u'title': u'Document 1',
+                     u'token': u'title'}]
+>>>>>>> master
             })
+
+    def tearDown(self):
+        gsm = getGlobalSiteManager()
+        gsm.unregisterUtility(provided=IVocabularyFactory,
+                              name='plone.restapi.tests.test_vocabulary')