+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+
 3.0.0 (unreleased)
 ------------------
 
@@ -40,6 +41,121 @@ Bugfixes:
 
 - Avoid calculating batch links for catalog results twice.
   [davisagli]
+
+
+3.2.2 (unreleased)
+------------------
+
+Bugfixes:
+
+- Do not include HTTP examples using data_files anymore, but move them below
+  src/plone/restapi instead and use package_data to include them.
+  [lgraf]
+
+- Rename Dexterity content before adding it to a container.
+  [buchi]
+
+- Avoid hard dependency on Archetypes introduced in 3.0.0.
+  This fixes `issue 570 <https://github.com/plone/plone.restapi/issues/570>`_.
+  [buchi]
+
+- Make setup.py require plone.behavior >= 1.1. This fixes #575.
+  [timo]
+
+
+3.2.1 (2018-06-28)
+------------------
+
+Bugfixes:
+
+- Require plone.schema >= 1.2.0 in setup.py for new tiles endpoint.
+  [timo]
+
+3.2.0 (2018-06-28)
+------------------
+
+New Features:
+
+- Add tiles endpoint for getting all available content tiles and its JSONSchema.
+  [sneridagh]
+
+- Add a tiles behavior to support the new tiles implementation for plone.restapi.
+  [sneridagh]
+
+- Make sure to include HTTP examples in installed egg, so test_documentation
+  tests also work against a installed release of plone.restapi.
+  [lgraf]
+
+
+3.1.0 (2018-06-27)
+------------------
+
+New Features:
+
+- Plone 5.2 compatibility.
+  [sunew, davisagli, timo]
+
+
+3.0.0 (2018-06-26)
+------------------
+
+Breaking Changes:
+
+- Fix object creation events. Before this fix, creation events were fired on
+  empty not yet deserialized objects. Also a modified event was fired after
+  deserializing e newly created object.
+  Custom content deserializers now must handle the `create` keyword argument,
+  which determines if deserialization is performed during object creation or
+  while updating an object.
+  [buchi]
+
+- Include translated role titles in `@sharing` GET.
+  [lgraf]
+
+- Image URLs are now created using the cache optimized way. Fixes #494.
+  [erral]
+
+
+2.2.1 (2018-06-25)
+------------------
+
+Bugfixes:
+
+- Fix ReST on PyPi.
+  [timo]
+
+
+2.2.0 (2018-06-25)
+------------------
+
+New Features:
+
+- Document the use of the `Accept-Language` HTTP header.
+  [erral]
+
+- Translate FTI titles on `@types` endpoint. Fixes #337.
+  [erral]
+
+- Translate action name, workflow state and transition names in @history endpoint.
+  [erral]
+
+- Enhance `@workflow` endpoint to support applying transitions to all contained
+  items and to set effective and expiration dates.
+  [buchi]
+
+Bugfixes:
+
+- Make sure DX DefaultFieldDeserializer validates field values.
+  [lgraf]
+
+- Reindex AT content on PATCH. This fixes `issue 531 <https://github.com/plone/plone.restapi/issues/531>`_.
+  [buchi]
+
+- Fix change password on Plone 5.2
+  [sunew]
+
+- Plone 5.2 compatible tests.
+  [sunew]
 
 
 2.1.0 (2018-06-23)